pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-tasks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "browser-use>=0.5.7",
    "google-genai>=1.28.0",
    "loguru>=0.7.3",
    "paapi5-python-sdk",
    "python-dotenv>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
paapi5-python-sdk = { path = "paapi5-python-sdk-example/paapi5-python-sdk-example" }

[project.scripts]
post = "src.main:main"
